rename files for consistency
