+++ finch/processes/wps_ensemble_indices_bbox.py
